requirements.txt
streamlit
python-docx
gspread
oauth2client